+++ src/lib/components/forms/form/example/form-fields.ts
@@ -135,14 +135,18 @@ export const formFields: TerraKeyValueInterface<TerraFormFieldInterface> = {
         options: {
             name: 'Textarea',
             required: true,
-            maxRows: 2
+            maxRows: 2,
+            maxLength: 50,
+            minLength: 2
         }
     },
     text: {
         type: 'text',
         options: {
             name: 'Text',
-            required: true
+            required: true,
+            maxLength: 10,
+            minLength: 2
         }
     },
     email: {